requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.7.2
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic[email]==2.7.1
httpx==0.27.0
stripe==9.5.0
pytesseract==0.3.10
opencv-python-headless==4.9.0.80
Pillow==10.3.0
numpy==1.26.4
slowapi==0.1.9
aiofiles==23.2.1